Fixed a stupid bug resulting of vector resizing.
Explanation:
The parser gets a pointer to the token value.
The token value itself is stored in a vector.
If the vector gets resized, the pointer maintained
by the parser dangles. In the end the parse result
is broken.

Solution:
Reserve 32 slots for tokens. This is more than engough
for line parsing. Abort if tokens vector is full.